requirements.txt
telethon==1.24.0
openai==1.3.5
requests==2.31.0
python-dotenv==1.0.0
Pillow==10.0.1
uvloop==0.19.0
flask==2.3.3
flask-cors==4.0.0
flask-jwt-extended
werkzeug
